requirements.txt
matplotlib
pre-commit
pytorch-lightning
pytorch-lightning-bolts
pyyaml
torch
torchvision
